Add username storing
The last logged in user’s ID gets stored in the session (until the tab
gets closed) or permanently in the local storage of the browser. When
the page is loaded, the website sends a request to the server asking
for the username and automatically fills the login field with it.