Added follow feature
This feature enables users to follow other users and see their content.
It still needs work, such as not allowing a user to follow someone
he/she already follows (right now the user he/she wants to follow is
just not added to the list of friends, but the page shouldn’t even
display the follow button).